+++ acceptance/tests/controller/controller_test.go
@@ -28,7 +28,9 @@ const (
 
 func TestController(t *testing.T) {
 	cfg := suite.Config()
-
+	if cfg.EnableCNI {
+		t.Skipf("skipping because -enable-cni is set and controller is already tested with regular tproxy")
+	}
 	cases := []struct {
 		secure      bool
 		autoEncrypt bool