+++ control-plane/connect-inject/container_init_test.go
@@ -228,48 +228,24 @@ consul-k8s-control-plane connect-init -pod-name=${POD_NAME} -pod-namespace=${POD
 func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 	cases := map[string]struct {
 		globalEnabled          bool
+		cniEnabled             bool
 		annotations            map[string]string
 		expectedContainsCmd    string
 		expectedNotContainsCmd string
 		namespaceLabel         map[string]string
 	}{
-		"enabled globally, ns not set, annotation not provided": {
+		"enabled globally, ns not set, annotation not provided, cni disabled": {
 			true,
-			nil,
-			`/consul/connect-inject/consul connect redirect-traffic \
-  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
-  -proxy-uid=5995`,
-			"",
-			nil,
-		},
-		"enabled globally, ns not set, annotation is false": {
-			true,
-			map[string]string{keyTransparentProxy: "false"},
-			"",
-			`/consul/connect-inject/consul connect redirect-traffic \
-  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
-  -proxy-uid=5995`,
-			nil,
-		},
-		"enabled globally, ns not set, annotation is true": {
-			true,
-			map[string]string{keyTransparentProxy: "true"},
-			`/consul/connect-inject/consul connect redirect-traffic \
-  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
-  -proxy-uid=5995`,
-			"",
-			nil,
-		},
-		"disabled globally, ns not set, annotation not provided": {
 			false,
 			nil,
-			"",
 			`/consul/connect-inject/consul connect redirect-traffic \
   -proxy-id="$(cat /consul/connect-inject/proxyid)" \
   -proxy-uid=5995`,
+			"",
 			nil,
 		},
-		"disabled globally, ns not set, annotation is false": {
+		"enabled globally, ns not set, annotation is false, cni disabled": {
+			true,
 			false,
 			map[string]string{keyTransparentProxy: "false"},
 			"",
@@ -278,7 +254,8 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
   -proxy-uid=5995`,
 			nil,
 		},
-		"disabled globally, ns not set, annotation is true": {
+		"enabled globally, ns not set, annotation is true, cni disabled": {
+			true,
 			false,
 			map[string]string{keyTransparentProxy: "true"},
 			`/consul/connect-inject/consul connect redirect-traffic \
@@ -287,8 +264,39 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			nil,
 		},
-		"exclude-inbound-ports, ns is not set, annotation is provided": {
+		"disabled globally, ns not set, annotation not provided, cni disabled": {
+			false,
+			false,
+			nil,
+			"",
+			`/consul/connect-inject/consul connect redirect-traffic \
+  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
+  -proxy-uid=5995`,
+			nil,
+		},
+		"disabled globally, ns not set, annotation is false, cni disabled": {
+			false,
+			false,
+			map[string]string{keyTransparentProxy: "false"},
+			"",
+			`/consul/connect-inject/consul connect redirect-traffic \
+  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
+  -proxy-uid=5995`,
+			nil,
+		},
+		"disabled globally, ns not set, annotation is true, cni disabled": {
+			false,
+			false,
+			map[string]string{keyTransparentProxy: "true"},
+			`/consul/connect-inject/consul connect redirect-traffic \
+  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
+  -proxy-uid=5995`,
+			"",
+			nil,
+		},
+		"exclude-inbound-ports, ns is not set, annotation is provided, cni disabled": {
 			true,
+			false,
 			map[string]string{
 				keyTransparentProxy:                 "true",
 				annotationTProxyExcludeInboundPorts: "9090,9091",
@@ -301,8 +309,9 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			nil,
 		},
-		"exclude-outbound-ports, ns is not set, annotation is provided": {
+		"exclude-outbound-ports, ns is not set, annotation is provided, cni disabled": {
 			true,
+			false,
 			map[string]string{
 				keyTransparentProxy:                  "true",
 				annotationTProxyExcludeOutboundPorts: "9090,9091",
@@ -315,8 +324,9 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			nil,
 		},
-		"exclude-outbound-cidrs annotation is provided": {
+		"exclude-outbound-cidrs annotation is provided, cni disabled": {
 			true,
+			false,
 			map[string]string{
 				keyTransparentProxy:                  "true",
 				annotationTProxyExcludeOutboundCIDRs: "1.1.1.1,2.2.2.2/24",
@@ -329,8 +339,9 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			nil,
 		},
-		"exclude-uids annotation is provided, ns is not set": {
+		"exclude-uids annotation is provided, ns is not set, cni disabled": {
 			true,
+			false,
 			map[string]string{
 				keyTransparentProxy:         "true",
 				annotationTProxyExcludeUIDs: "6000,7000",
@@ -343,7 +354,8 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			nil,
 		},
-		"disabled globally, ns enabled, annotation not set": {
+		"disabled globally, ns enabled, annotation not set, cni disabled": {
+			false,
 			false,
 			nil,
 			`/consul/connect-inject/consul connect redirect-traffic \
@@ -352,8 +364,9 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 			"",
 			map[string]string{keyTransparentProxy: "true"},
 		},
-		"enabled globally, ns disabled, annotation not set": {
+		"enabled globally, ns disabled, annotation not set, cni disabled": {
 			true,
+			false,
 			nil,
 			"",
 			`/consul/connect-inject/consul connect redirect-traffic \
@@ -361,24 +374,56 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
   -proxy-uid=5995`,
 			map[string]string{keyTransparentProxy: "false"},
 		},
+		"disabled globally, ns enabled, annotation not set, cni enabled": {
+			false,
+			true,
+			nil,
+			"",
+			`/consul/connect-inject/consul connect redirect-traffic \
+  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
+  -proxy-uid=5995`,
+			map[string]string{keyTransparentProxy: "true"},
+		},
+
+		"enabled globally, ns not set, annotation not set, cni enabled": {
+			true,
+			true,
+			nil,
+			"",
+			`/consul/connect-inject/consul connect redirect-traffic \
+  -proxy-id="$(cat /consul/connect-inject/proxyid)" \
+  -proxy-uid=5995`,
+			nil,
+		},
 	}
 	for name, c := range cases {
 		t.Run(name, func(t *testing.T) {
 			w := MeshWebhook{
 				EnableTransparentProxy: c.globalEnabled,
 				ConsulAPITimeout:       5 * time.Second,
+				EnableCNI:              c.cniEnabled,
 			}
 			pod := minimal()
 			pod.Annotations = c.annotations
 
-			expectedSecurityContext := &corev1.SecurityContext{
-				RunAsUser:  pointer.Int64(0),
-				RunAsGroup: pointer.Int64(0),
-				Privileged: pointer.Bool(true),
-				Capabilities: &corev1.Capabilities{
+			expectedSecurityContext := &corev1.SecurityContext{}
+			if !c.cniEnabled {
+				expectedSecurityContext.RunAsUser = pointer.Int64(0)
+				expectedSecurityContext.RunAsGroup = pointer.Int64(0)
+				expectedSecurityContext.RunAsNonRoot = pointer.Bool(false)
+				expectedSecurityContext.Privileged = pointer.Bool(true)
+				expectedSecurityContext.Capabilities = &corev1.Capabilities{
 					Add: []corev1.Capability{netAdminCapability},
-				},
-				RunAsNonRoot: pointer.Bool(false),
+				}
+			} else {
+
+				expectedSecurityContext.RunAsUser = pointer.Int64(initContainersUserAndGroupID)
+				expectedSecurityContext.RunAsGroup = pointer.Int64(initContainersUserAndGroupID)
+				expectedSecurityContext.RunAsNonRoot = pointer.Bool(true)
+				expectedSecurityContext.Privileged = pointer.Bool(false)
+				expectedSecurityContext.Capabilities = &corev1.Capabilities{
+					Drop: []corev1.Capability{"ALL"},
+				}
 			}
 			ns := testNS
 			ns.Labels = c.namespaceLabel
@@ -390,7 +435,11 @@ func TestHandlerContainerInit_transparentProxy(t *testing.T) {
 				require.Equal(t, expectedSecurityContext, container.SecurityContext)
 				require.Contains(t, actualCmd, c.expectedContainsCmd)
 			} else {
-				require.Nil(t, container.SecurityContext)
+				if !c.cniEnabled {
+					require.Nil(t, container.SecurityContext)
+				} else {
+					require.Equal(t, expectedSecurityContext, container.SecurityContext)
+				}
 				require.NotContains(t, actualCmd, c.expectedNotContainsCmd)
 			}
 		})
@@ -917,7 +966,8 @@ func TestHandlerContainerInit_Multiport(t *testing.T) {
 					serviceName:  "web-admin",
 				},
 			},
-			[]string{`/bin/sh -ec 
+			[]string{
+				`/bin/sh -ec 
 export CONSUL_HTTP_ADDR="${HOST_IP}:8500"
 export CONSUL_GRPC_ADDR="${HOST_IP}:8502"
 consul-k8s-control-plane connect-init -pod-name=${POD_NAME} -pod-namespace=${POD_NAMESPACE} \
@@ -968,7 +1018,8 @@ consul-k8s-control-plane connect-init -pod-name=${POD_NAME} -pod-namespace=${POD
 					serviceName:  "web-admin",
 				},
 			},
-			[]string{`/bin/sh -ec 
+			[]string{
+				`/bin/sh -ec 
 export CONSUL_HTTP_ADDR="${HOST_IP}:8500"
 export CONSUL_GRPC_ADDR="${HOST_IP}:8502"
 consul-k8s-control-plane connect-init -pod-name=${POD_NAME} -pod-namespace=${POD_NAMESPACE} \
@@ -1167,8 +1218,8 @@ func TestHandlerInitCopyContainer(t *testing.T) {
 				require.Nil(t, container.SecurityContext)
 			} else {
 				expectedSecurityContext := &corev1.SecurityContext{
-					RunAsUser:              pointer.Int64(copyContainerUserAndGroupID),
-					RunAsGroup:             pointer.Int64(copyContainerUserAndGroupID),
+					RunAsUser:              pointer.Int64(initContainersUserAndGroupID),
+					RunAsGroup:             pointer.Int64(initContainersUserAndGroupID),
 					RunAsNonRoot:           pointer.Bool(true),
 					ReadOnlyRootFilesystem: pointer.Bool(true),
 				}