+++ control-plane/connect-inject/mesh_webhook.go
@@ -25,11 +25,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
-var (
-	// kubeSystemNamespaces is a set of namespaces that are considered
-	// "system" level namespaces and are always skipped (never injected).
-	kubeSystemNamespaces = mapset.NewSetWith(metav1.NamespaceSystem, metav1.NamespacePublic)
-)
+// kubeSystemNamespaces is a set of namespaces that are considered
+// "system" level namespaces and are always skipped (never injected).
+var kubeSystemNamespaces = mapset.NewSetWith(metav1.NamespaceSystem, metav1.NamespacePublic)
 
 // Webhook is the HTTP meshWebhook for admission webhooks.
 type MeshWebhook struct {
@@ -135,6 +133,11 @@ type MeshWebhook struct {
 	// This means that the injected init container will apply traffic redirection rules
 	// so that all traffic will go through the Envoy proxy.
 	EnableTransparentProxy bool
+
+	// EnableCNI enables the CNI plugin and prevents the connect-inject init container
+	// from running the consul redirect-traffic command as the CNI plugin handles traffic
+	// redirection
+	EnableCNI bool
 
 	// TProxyOverwriteProbes controls whether the webhook should mutate pod's HTTP probes
 	// to point them to the Envoy proxy.
@@ -362,6 +365,18 @@ func (w *MeshWebhook) Handle(ctx context.Context, req admission.Request) admissi
 	// and does not need to be checked for being a nil value.
 	pod.Annotations[keyInjectStatus] = injected
 
+	tproxyEnabled, err := transparentProxyEnabled(*ns, pod, w.EnableTransparentProxy)
+	if err != nil {
+		w.Log.Error(err, "error determining if transparent proxy is enabled", "request name", req.Name)
+		return admission.Errored(http.StatusInternalServerError, fmt.Errorf("error determining if transparent proxy is enabled: %s", err))
+	}
+
+	// Add an annotation to the pod sets transparent-proxy-status to enabled or disabled. Used by the CNI plugin
+	// to determine if it should traffic redirect or not
+	if tproxyEnabled {
+		pod.Annotations[keyTransparentProxyStatus] = enabled
+	}
+
 	// Add annotations for metrics.
 	if err = w.prometheusAnnotations(&pod); err != nil {
 		w.Log.Error(err, "error configuring prometheus annotations", "request name", req.Name)
@@ -387,6 +402,16 @@ func (w *MeshWebhook) Handle(ctx context.Context, req admission.Request) admissi
 	if err != nil {
 		w.Log.Error(err, "error overwriting readiness or liveness probes", "request name", req.Name)
 		return admission.Errored(http.StatusInternalServerError, fmt.Errorf("error overwriting readiness or liveness probes: %s", err))
+	}
+
+	// When CNI and tproxy are enabled, we add an annotation to the pod that contains the iptables config so that the CNI
+	// plugin can apply redirect traffic rules on the pod.
+	if w.EnableCNI && tproxyEnabled {
+		if err := w.addRedirectTrafficConfigAnnotation(&pod, *ns); err != nil {
+			// todo: update this error message
+			w.Log.Error(err, "error configuring annotation for CNI traffic redirection", "request name", req.Name)
+			return admission.Errored(http.StatusInternalServerError, fmt.Errorf("error configuring annotation for CNI traffic redirection: %s", err))
+		}
 	}
 
 	// Marshall the pod into JSON after it has the desired envs, annotations, labels,